only copy solarized palette if missing

+++ solarized/install.sh
@@ -1,1 +1,4 @@
-wget https://github.com/altercation/solarized/blob/master/apple-colorpalette-solarized/solarized.clr?raw=true -O ~/Library/Colors/Solarized.clr
+if [ ! -f ~/Library/Colors/Solarized.clr ]
+then
+    wget https://github.com/altercation/solarized/blob/master/apple-colorpalette-solarized/solarized.clr?raw=true -O ~/Library/Colors/Solarized.clr
+fi
